+++ utils/helpers.py
@@ -599,7 +599,7 @@ def validate_facebook_credentials(config_obj) -> tuple[bool, str]:
     return True, ""
 
 
-CURRENT_VERSION = "v5.0.28"
+CURRENT_VERSION = "v5.0.29"
 
 
 def get_current_version() -> str: